🧹 [CleanUp] Created own file for themeState
